refactor(stage1): move citation rendering to woeai.publications.rendering
Candidate A, slice 4. Moves the rendered_entry pipeline and its emphasis
markers out of the script:

- woeai/publications/rendering.py: bold_group_leader, bold_journal_title,
  bold_metric_values, rendered_entry (the [N] ... citation composer).
- Depends on textutils (rst_escape, strip_bib_html, METRIC_LABELS) and
  authors (mark_student_first_author, mark_corresponding_authors) - the
  earlier slices.
- Script re-exports all four so tests (self.updater.rendered_entry) unchanged.

Stage 1 migration of the pure-computation core is now complete: the script
keeps only Zotero HTTP fetch, research-map loading, anchor logic, page
assembly, snapshot output, and CLI. 54 tests green.
